requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
Django==3.1.3
django-tastypie==0.14.3
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
python-mimeparse==1.6.0
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.1
